+++ gazon-payment/src/main/java/com/gitlab/GazonPaymentApplication.java
@@ -1,0 +1,12 @@
+package com.gitlab;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GazonPaymentApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GazonPaymentApplication.class, args);
+	}
+}