+++ gazon-service/src/test/java/com/gitlab/service/PaymentServiceTest.java
@@ -1,6 +1,13 @@
 package com.gitlab.service;
 
+import com.gitlab.client.PaymentClient;
+import com.gitlab.dto.OrderDto;
+import com.gitlab.dto.PaymentDto;
+import com.gitlab.enums.OrderStatus;
+import com.gitlab.enums.PaymentStatus;
+import com.gitlab.mapper.PaymentMapper;
 import com.gitlab.model.Payment;
+import com.gitlab.model.User;
 import com.gitlab.repository.PaymentRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -8,6 +15,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,6 +29,19 @@ class PaymentServiceTest {
 
     @Mock
     private PaymentRepository paymentRepository;
+
+    @Mock
+    private UserService userService;
+
+    @Mock
+    private PaymentMapper paymentMapper;
+
+    @Mock
+    private PaymentClient paymentClient;
+
+    @Mock
+    private OrderService orderService;
+
     @InjectMocks
     private PaymentService paymentService;
 
@@ -52,6 +74,48 @@ class PaymentServiceTest {
         Payment actualResult = paymentService.save(expectedResult);
 
         assertEquals(expectedResult, actualResult);
+    }
+
+    @Test
+    public void testSaveDto_Success() {
+        User user = new User();
+        user.setId(123L);
+
+        PaymentDto paymentDto = new PaymentDto();
+        paymentDto.setUserId(123L);
+        paymentDto.setPaymentStatus(PaymentStatus.NOT_PAID);
+
+        Payment payment = new Payment();
+
+        Payment savedPayment = new Payment();
+        PaymentDto savedPaymentDto = new PaymentDto();
+        savedPaymentDto.setPaymentStatus(PaymentStatus.PAID);
+        savedPaymentDto.setOrderId(1L);
+
+        PaymentDto paymentDtoResponse = new PaymentDto();
+        paymentDtoResponse.setPaymentStatus(PaymentStatus.PAID);
+        paymentDtoResponse.setOrderId(1L);
+
+        ResponseEntity<PaymentDto> paymentDtoResponseEntity = new ResponseEntity<>(paymentDtoResponse, HttpStatus.OK);
+
+        OrderDto orderDto = new OrderDto();
+        orderDto.setOrderStatus(OrderStatus.NOT_PAID);
+
+        when(userService.getAuthenticatedUser()).thenReturn(user);
+        when(paymentMapper.toEntity(paymentDto)).thenReturn(payment);
+        when(paymentRepository.save(payment)).thenReturn(savedPayment);
+        Payment savedDtoResponseEntity = new Payment();
+        savedDtoResponseEntity.setId(1L);
+        when(paymentRepository.save(paymentMapper.toEntity(paymentDtoResponse))).thenReturn(savedDtoResponseEntity);
+        when(paymentMapper.toDto(savedPayment)).thenReturn(savedPaymentDto);
+        when(paymentClient.makePayment(any(PaymentDto.class))).thenReturn(paymentDtoResponseEntity);
+        when(orderService.findByIdDto(1L)).thenReturn(Optional.of(orderDto));
+
+        PaymentDto result = paymentService.saveDto(paymentDto);
+
+        assertEquals(PaymentStatus.PAID, result.getPaymentStatus());
+        verify(orderService).saveDto(orderDto);
+        assertEquals(OrderStatus.PAID, orderDto.getOrderStatus());
     }
 
     @Test